# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for FemmeFlow dropshipping store
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  costPrice: number;
  sellingPrice: number;
  profitMargin: number;
  category: string;
  imageUrl: string;
  stockCount: number;
  rating: number;
  reviewCount: number;
  isTrending: boolean;
  viralReason?: string;
  targetAudience?: string;
  supplierUrl?: string;
  tags?: string[];
}

export interface Review {
  id: number;
  productId: number;
  authorName: string;
  rating: number;
  comment: string;
  createdAt: string;
  avatarUrl?: string;
}

export interface CreateReviewBody {
  productId: number;
  authorName: string;
  rating: number;
  comment: string;
}

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  productImage: string;
  quantity: number;
  price: number;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartBody {
  sessionId: string;
  productId: number;
  quantity: number;
}

export interface CreateOrderBody {
  sessionId: string;
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
  paymentMethod: string;
}

export interface Order {
  id: number;
  sessionId: string;
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
  paymentMethod: string;
  status: string;
  total: number;
  createdAt: string;
  items: CartItem[];
}

export interface CatalogSummary {
  totalProducts: number;
  totalCategories: number;
  trendingCount: number;
  topCategory: string;
  avgRating: number;
  totalReviews: number;
}

export type ListProductsParams = {
  category?: string;
  limit?: number;
  offset?: number;
};

export type ListReviewsParams = {
  productId: number;
};

export type GetCartParams = {
  sessionId: string;
};