# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for FemmeFlow dropshipping store
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog
  - name: cart
    description: Shopping cart
  - name: orders
    description: Order management
  - name: reviews
    description: Product reviews
  - name: catalog
    description: Catalog summaries and trending data

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  /products/trending:
    get:
      operationId: getTrendingProducts
      tags: [catalog]
      summary: Get trending/viral products
      responses:
        "200":
          description: Trending products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /catalog/summary:
    get:
      operationId: getCatalogSummary
      tags: [catalog]
      summary: Get store catalog summary stats
      responses:
        "200":
          description: Catalog summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogSummary"

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: productId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/item/{itemId}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        costPrice:
          type: number
        sellingPrice:
          type: number
        profitMargin:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        stockCount:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        isTrending:
          type: boolean
        viralReason:
          type: string
        targetAudience:
          type: string
        supplierUrl:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - description
        - costPrice
        - sellingPrice
        - profitMargin
        - category
        - imageUrl
        - stockCount
        - rating
        - reviewCount
        - isTrending

    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        authorName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        avatarUrl:
          type: string
      required:
        - id
        - productId
        - authorName
        - rating
        - comment
        - createdAt

    CreateReviewBody:
      type: object
      properties:
        productId:
          type: integer
        authorName:
          type: string
        rating:
          type: integer
        comment:
          type: string
      required:
        - productId
        - authorName
        - rating
        - comment

    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - id
        - productId
        - productName
        - productImage
        - quantity
        - price

    Cart:
      type: object
      properties:
        sessionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer
      required:
        - sessionId
        - items
        - total
        - itemCount

    AddToCartBody:
      type: object
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - sessionId
        - productId
        - quantity

    CreateOrderBody:
      type: object
      properties:
        sessionId:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        shippingAddress:
          type: string
        paymentMethod:
          type: string
      required:
        - sessionId
        - customerName
        - customerEmail
        - shippingAddress
        - paymentMethod

    Order:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        shippingAddress:
          type: string
        paymentMethod:
          type: string
        status:
          type: string
        total:
          type: number
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
      required:
        - id
        - sessionId
        - customerName
        - customerEmail
        - shippingAddress
        - paymentMethod
        - status
        - total
        - createdAt
        - items

    CatalogSummary:
      type: object
      properties:
        totalProducts:
          type: integer
        totalCategories:
          type: integer
        trendingCount:
          type: integer
        topCategory:
          type: string
        avgRating:
          type: number
        totalReviews:
          type: integer
      required:
        - totalProducts
        - totalCategories
        - trendingCount
        - topCategory
        - avgRating
        - totalReviews