# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for FemmeFlow dropshipping store
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all products
 */
export const listProductsQueryLimitDefault = 20;
export const listProductsQueryOffsetDefault = 0;

export const ListProductsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  limit: zod.coerce.number().default(listProductsQueryLimitDefault),
  offset: zod.coerce.number().default(listProductsQueryOffsetDefault),
});

export const ListProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  costPrice: zod.number(),
  sellingPrice: zod.number(),
  profitMargin: zod.number(),
  category: zod.string(),
  imageUrl: zod.string(),
  stockCount: zod.number(),
  rating: zod.number(),
  reviewCount: zod.number(),
  isTrending: zod.boolean(),
  viralReason: zod.string().optional(),
  targetAudience: zod.string().optional(),
  supplierUrl: zod.string().optional(),
  tags: zod.array(zod.string()).optional(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Get a product by ID
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  costPrice: zod.number(),
  sellingPrice: zod.number(),
  profitMargin: zod.number(),
  category: zod.string(),
  imageUrl: zod.string(),
  stockCount: zod.number(),
  rating: zod.number(),
  reviewCount: zod.number(),
  isTrending: zod.boolean(),
  viralReason: zod.string().optional(),
  targetAudience: zod.string().optional(),
  supplierUrl: zod.string().optional(),
  tags: zod.array(zod.string()).optional(),
});

/**
 * @summary Get trending/viral products
 */
export const GetTrendingProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  costPrice: zod.number(),
  sellingPrice: zod.number(),
  profitMargin: zod.number(),
  category: zod.string(),
  imageUrl: zod.string(),
  stockCount: zod.number(),
  rating: zod.number(),
  reviewCount: zod.number(),
  isTrending: zod.boolean(),
  viralReason: zod.string().optional(),
  targetAudience: zod.string().optional(),
  supplierUrl: zod.string().optional(),
  tags: zod.array(zod.string()).optional(),
});
export const GetTrendingProductsResponse = zod.array(
  GetTrendingProductsResponseItem,
);

/**
 * @summary Get store catalog summary stats
 */
export const GetCatalogSummaryResponse = zod.object({
  totalProducts: zod.number(),
  totalCategories: zod.number(),
  trendingCount: zod.number(),
  topCategory: zod.string(),
  avgRating: zod.number(),
  totalReviews: zod.number(),
});

/**
 * @summary List reviews for a product
 */
export const ListReviewsQueryParams = zod.object({
  productId: zod.coerce.number(),
});

export const ListReviewsResponseItem = zod.object({
  id: zod.number(),
  productId: zod.number(),
  authorName: zod.string(),
  rating: zod.number(),
  comment: zod.string(),
  createdAt: zod.coerce.date(),
  avatarUrl: zod.string().optional(),
});
export const ListReviewsResponse = zod.array(ListReviewsResponseItem);

/**
 * @summary Submit a review
 */
export const CreateReviewBody = zod.object({
  productId: zod.number(),
  authorName: zod.string(),
  rating: zod.number(),
  comment: zod.string(),
});

/**
 * @summary Get current cart
 */
export const GetCartQueryParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const GetCartResponse = zod.object({
  sessionId: zod.string(),
  items: zod.array(
    zod.object({
      id: zod.number(),
      productId: zod.number(),
      productName: zod.string(),
      productImage: zod.string(),
      quantity: zod.number(),
      price: zod.number(),
    }),
  ),
  total: zod.number(),
  itemCount: zod.number(),
});

/**
 * @summary Add item to cart
 */
export const AddToCartBody = zod.object({
  sessionId: zod.string(),
  productId: zod.number(),
  quantity: zod.number(),
});

export const AddToCartResponse = zod.object({
  sessionId: zod.string(),
  items: zod.array(
    zod.object({
      id: zod.number(),
      productId: zod.number(),
      productName: zod.string(),
      productImage: zod.string(),
      quantity: zod.number(),
      price: zod.number(),
    }),
  ),
  total: zod.number(),
  itemCount: zod.number(),
});

/**
 * @summary Remove item from cart
 */
export const RemoveFromCartParams = zod.object({
  itemId: zod.coerce.number(),
});

export const RemoveFromCartResponse = zod.object({
  sessionId: zod.string(),
  items: zod.array(
    zod.object({
      id: zod.number(),
      productId: zod.number(),
      productName: zod.string(),
      productImage: zod.string(),
      quantity: zod.number(),
      price: zod.number(),
    }),
  ),
  total: zod.number(),
  itemCount: zod.number(),
});

/**
 * @summary Place an order
 */
export const CreateOrderBody = zod.object({
  sessionId: zod.string(),
  customerName: zod.string(),
  customerEmail: zod.string(),
  shippingAddress: zod.string(),
  paymentMethod: zod.string(),
});