# lib/api-zod/src/generated/types/catalogSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for FemmeFlow dropshipping store
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogSummary {
  totalProducts: number;
  totalCategories: number;
  trendingCount: number;
  topCategory: string;
  avgRating: number;
  totalReviews: number;
}