# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for FemmeFlow dropshipping store
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from "./cartItem";

export interface Order {
  id: number;
  sessionId: string;
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
  paymentMethod: string;
  status: string;
  total: number;
  createdAt: Date;
  items: CartItem[];
}