# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for FemmeFlow dropshipping store
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  costPrice: number;
  sellingPrice: number;
  profitMargin: number;
  category: string;
  imageUrl: string;
  stockCount: number;
  rating: number;
  reviewCount: number;
  isTrending: boolean;
  viralReason?: string;
  targetAudience?: string;
  supplierUrl?: string;
  tags?: string[];
}