# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for FemmeFlow dropshipping store
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  productId: number;
  authorName: string;
  rating: number;
  comment: string;
  createdAt: Date;
  avatarUrl?: string;
}